requirements.txt
streamlit==1.47.0
fastapi==0.115.14
pydantic==2.11.7
uvicorn==0.35.0
mysql-connector-python==9.3.0
pandas==2.2.3
requests==2.32.3
pytest==8.4.1
fpdf==1.7.2
python-dotenv~=0.21.0
